requirements.txt
openai-whisper
fastapi
uvicorn
gevent-websocket
python-multipart
transformers
huggingface_hub
torch==2.0.1
librosa==0.10.0
torchvision==0.15.2 
torchaudio==2.0.2
tqdm
numpy

txtsplit
cached_path
python-crfsuite==0.9.10
num2words==0.5.12
unidic_lite==1.0.8
unidic==1.1.0
pykakasi==2.2.1
fugashi==1.3.0
g2p_en==2.1.0
anyascii==0.3.2
jamo==0.4.1
gruut[de,es,fr]==2.2.3
g2pkk>=0.1.1
pydub==0.25.1
eng_to_ipa==0.0.2
inflect==7.0.0
unidecode==1.3.7
pypinyin==0.50.0
cn2an==0.5.22
jieba==0.42.1
langid==1.1.6
